[TF FE] Move hash_table header to TF FE public
It is needed to duplicate the current implementation in ov tokenizers
and extend it to support string type case based on OpExtension
